Split the stimulus controllers into their own file
The stimulus application was being initialized in Argo.initialize, which is called on every page load.  This is not ideal, especially when using turbo or turbolinks.  This also separates the stimulus from the non-stimulus part of the application
